+++ client-config/src/main/java/com/networknt/client/OAuthSignConfig.java
@@ -1,35 +1,28 @@
 package com.networknt.client;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.networknt.config.schema.BooleanField;
 import com.networknt.config.schema.IntegerField;
 import com.networknt.config.schema.ObjectField;
 import com.networknt.config.schema.StringField;
 
-public class OauthSignConfig {
-
-    public static final String SERVER_URL = "server_url";
-    public static final String PROXY_HOST = "proxyHost";
-    public static final String PROXY_PORT = "proxyPort";
-    public static final String SERVICE_ID = "serviceId";
-    public static final String URI = "uri";
-    public static final String TIMEOUT = "timeout";
-    public static final String ENABLE_HTTP_2 = "enableHttp2";
-    public static final String CLIENT_ID = "client_id";
-    public static final String CLIENT_SECRET = "client_secret";
-    public static final String KEY = "key";
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class OAuthSignConfig {
 
     @StringField(
-            configFieldName = SERVER_URL,
+            configFieldName = ClientConfig.SERVER_URL,
             externalizedKeyName = "signServerUrl",
             externalized = true,
             description = "token server url. The default port number for token service is 6882. " +
                     "If this url exists, it will be used.\n" +
                     "if it is not set, then a service lookup against serviceId will be taken to discover an instance."
     )
-    private String server_url;
+    @JsonProperty(ClientConfig.SERVER_URL)
+    private String serverUrl = null;
 
     @StringField(
-            configFieldName = PROXY_HOST,
+            configFieldName = ClientConfig.PROXY_HOST,
             externalizedKeyName = "signProxyHost",
             externalized = true,
             description = "For users who leverage SaaS OAuth 2.0 provider from lightapi.net or others in the public cloud\n" +
@@ -37,10 +30,11 @@ public class OauthSignConfig {
                     "proxyHost here for the HTTPS traffic. This option is only working with server_url and serviceId\n" +
                     "below should be commented out. OAuth 2.0 services cannot be discovered if a proxy server is used."
     )
-    private String proxyHost;
+    @JsonProperty(ClientConfig.PROXY_HOST)
+    private String proxyHost = null;
 
     @IntegerField(
-            configFieldName = PROXY_PORT,
+            configFieldName = ClientConfig.PROXY_PORT,
             externalizedKeyName = "signProxyPort",
             min = 0,
             max = 65535,
@@ -49,81 +43,89 @@ public class OauthSignConfig {
                     "a different port, please specify it here. If proxyHost is available and proxyPort is missing, then\n" +
                     "the default value 443 is going to be used for the HTTP connection."
     )
-    private int proxyPort;
+    @JsonProperty(ClientConfig.PROXY_PORT)
+    private Integer proxyPort = null;
 
     @StringField(
-            configFieldName = SERVICE_ID,
+            configFieldName = ClientConfig.SERVICE_ID,
             externalizedKeyName = "signServiceId",
             externalized = true,
             defaultValue = "com.networknt.oauth2-token-1.0.0",
             description = "token serviceId. If server_url doesn't exist, " +
                     "the serviceId will be used to lookup the token service."
     )
-    private String serviceId;
+    @JsonProperty(ClientConfig.SERVICE_ID)
+    private String serviceId = "com.networknt.oauth2-token-1.0.0";
 
     @StringField(
-            configFieldName = URI,
+            configFieldName = ClientConfig.URI,
             externalizedKeyName = "signUri",
             externalized = true,
             defaultValue = "/oauth2/token",
             description = "signing endpoint for the sign request"
     )
-    private String uri;
+    @JsonProperty(ClientConfig.URI)
+    private String uri = "/oauth2/sign";
 
     @IntegerField(
-            configFieldName = TIMEOUT,
+            configFieldName = ClientConfig.TIMEOUT,
             externalizedKeyName = "signTimeout",
             externalized = true,
-            defaultValue = 2000,
+            defaultValue = "2000",
             description = "timeout in milliseconds"
     )
-    private int timeout;
+    @JsonProperty(ClientConfig.TIMEOUT)
+    private Integer timeout = 2000;
 
     @BooleanField(
-            configFieldName = ENABLE_HTTP_2,
+            configFieldName = ClientConfig.ENABLE_HTTP_2,
             externalizedKeyName = "signEnableHttp2",
-            defaultValue = true,
+            defaultValue = "true",
             externalized = true,
             description = "set to true if the oauth2 provider supports HTTP/2"
     )
-    private boolean enableHttp2;
+    @JsonProperty(ClientConfig.ENABLE_HTTP_2)
+    private boolean enableHttp2 = true;
 
     @StringField(
-            configFieldName = CLIENT_ID,
+            configFieldName = ClientConfig.CLIENT_ID,
             externalizedKeyName = "signClientId",
             externalized = true,
             defaultValue = "f7d42348-c647-4efb-a52d-4c5787421e72",
             description = "client_id for client authentication"
     )
-    private char[] client_id;
+    @JsonProperty(ClientConfig.CLIENT_ID)
+    private char[] clientId =  "f7d42348-c647-4efb-a52d-4c5787421e72".toCharArray();
 
     @StringField(
-            configFieldName = CLIENT_SECRET,
+            configFieldName = ClientConfig.CLIENT_SECRET,
             externalizedKeyName = "signClientSecret",
             externalized = true,
             defaultValue = "f6h1FTI8Q3-7UScPZDzfXA",
             description = "client secret for client authentication and it can be encrypted here."
     )
-    private char[] client_secret;
+    @JsonProperty(ClientConfig.CLIENT_SECRET)
+    private char[] clientSecret = "f6h1FTI8Q3-7UScPZDzfXA".toCharArray();
 
     @ObjectField(
-            configFieldName = KEY,
+            configFieldName = ClientConfig.KEY,
             ref = OAuthSignKeyConfig.class,
             useSubObjectDefault = true,
             description = "the key distribution sever config for sign. " +
                     "It can be different then token key distribution server."
     )
-    OAuthSignKeyConfig key;
+    @JsonProperty(ClientConfig.KEY)
+    OAuthSignKeyConfig key = null;
 
-    public String getServer_url() {
-        return server_url;
+    public String getServerUrl() {
+        return serverUrl;
     }
 
     public String getProxyHost() {
         return proxyHost;
     }
 
-    public int getProxyPort() {
+    public Integer getProxyPort() {
         return proxyPort;
     }
 
@@ -135,7 +137,7 @@ public class OauthSignConfig {
         return uri;
     }
 
-    public int getTimeout() {
+    public Integer getTimeout() {
         return timeout;
     }
 
@@ -143,12 +145,12 @@ public class OauthSignConfig {
         return enableHttp2;
     }
 
-    public char[] getClient_id() {
-        return client_id;
+    public char[] getClientId() {
+        return clientId;
     }
 
-    public char[] getClient_secret() {
-        return client_secret;
+    public char[] getClientSecret() {
+        return clientSecret;
     }
 
     public OAuthSignKeyConfig getKey() {